refactor(host-application): enforce runtime cleanup invariants
Replace split runtime cleanup path options with a single RuntimeCleanupTarget invariant and thread it through startup/prerequisite/registry lifecycle flows.

Further decompose runtime registry into query, lifecycle, and cleanup modules while preserving runtime behavior and error propagation semantics.

Update integration tests to use the new cleanup target model.
